# back/api/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GPE Backend contract interface",
    "description": "This document provides an exhaustive list of all endpoints available on gpe backend.",
    "version": "0.1.1"
  },
  "servers": [
    {
      "url": "/",
      "description": ""
    },
    {
      "url": "http://localhost:8080",
      "description": "Running local dev server"
    }
  ],
...
  "paths": {
...
    "/barometer": {
      "post": {
        "operationId": "defineDailyBarometer",
        "tags": [
          "Barometer"
        ],
        "responses": {
          "200": {
            "description": "Barometer defined",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.barometer"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable entity, some validations failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.message"
                }
              }
            }
          }
        },
        "description": "Define the daily barometer for the current user. On barometer per day and barometer type. Score is the barometer evaluation. Post again barometer to update the daily value.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/input.barometer"
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          }
        ]
      },
      "get": {
        "operationId": "retrieveDailyBarometer",
        "tags": [
          "Barometer"
        ],
        "responses": {
          "200": {
            "description": "Barometers defined for current user for today",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.barometers"
                }
              }
            }
          }
        },
        "description": "Returns all barometers defined for the current user at this day.",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          }
        ]
      }
    },
    "/barometer/stats": {
      "get": {
        "operationId": "retrieveBarometerStats",
        "tags": [
          "Barometer"
        ],
        "responses": {
          "200": {
            "description": "Barometers stats for specified period and scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.barometer.stats"
                }
              }
            }
          }
        },
        "description": "[ADMIN] Returns all barometers stats, admin only.",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          },
          {
            "name": "scopeCategories",
            "in": "query",
            "description": "Get stats scoped by work categories",
            "required": false,
            "allowEmptyValue": false,
            "schema": {
              "type": "boolean",
              "example": false,
              "default": false
            }
          },
          {
            "name": "timeScope",
            "in": "query",
            "description": "Get stats grouped by time range period",
            "required": false,
            "allowEmptyValue": false,
            "schema": {
              "type": "string",
              "example": "year",
              "default": "year",
              "enum": [
                "year",
                "month",
                "day"
              ]
            }
          }
        ]
      }
    },
    "/notification": {
      "get": {
        "operationId": "getUnreadNotifications",
        "tags": [
          "Notification"
        ],
        "responses": {
          "200": {
            "description": "Unread notifications",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.notifications"
                }
              }
            }
          }
        },
        "description": "Returns all unread notifications for the current user.",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          },
          {
            "$ref": "#/components/parameters/filter.id"
          },
          {
            "$ref": "#/components/parameters/filter.createdAtBefore"
          },
          {
            "$ref": "#/components/parameters/filter.createdAtAfter"
          },
          {
            "$ref": "#/components/parameters/filter.updatedAtBefore"
          },
          {
            "$ref": "#/components/parameters/filter.updatedAtAfter"
          }
        ]
      }
    },
    "/notification/archive": {
      "get": {
        "operationId": "getNotifications",
        "tags": [
          "Notification"
        ],
        "responses": {
          "200": {
            "description": "All user notifications",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/output.notifications"
                }
              }
            }
          }
        },
        "description": "Returns all notifications for the current user.",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          },
          {
            "$ref": "#/components/parameters/filter.id"
          },
          {
            "$ref": "#/components/parameters/filter.createdAtBefore"
          },
          {
            "$ref": "#/components/parameters/filter.createdAtAfter"
          },
          {
            "$ref": "#/components/parameters/filter.updatedAtBefore"
          },
          {
            "$ref": "#/components/parameters/filter.updatedAtAfter"
          }
        ]
      }
    },
    "/notification/read": {
      "post": {
        "operationId": "setAllNotificationReadForCurrentUser",
        "tags": [
          "Notification"
        ],
        "responses": {
          "204": {
            "description": "Notifications reads"
          }
        },
        "description": "Set all current user notification to status read",
        "parameters": [
          {
            "$ref": "#/components/parameters/accessToken"
          }
        ]
      }
    },
...
  },
  "components": {
    "parameters": {
...
      "accessToken": {
        "name": "X-Access-Token",
        "in": "header",
        "description": "The access token",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
      "filter.id": {
        "name": "id[]",
        "in": "query",
        "description": "Resource identifiers",
        "required": false,
        "schema": {
          "type": "array",
          "items": {
            "type": "integer",
            "minimum": 1,
            "example": 1
          }
        },
        "explode": true
      },
      "filter.createdAtBefore": {
        "name": "createdAt[before]",
        "in": "query",
        "description": "Created at before the provided date",
        "required": false,
        "schema": {
          "type": "string",
          "format": "datetime"
        }
      },
      "filter.createdAtAfter": {
        "name": "createdAt[after]",
        "in": "query",
        "description": "Created after the provided date",
        "required": false,
        "schema": {
          "type": "string",
          "format": "datetime"
        }
      },
      "filter.updatedAtBefore": {
        "name": "updatedAt[before]",
        "in": "query",
        "description": "Updated at before the provided date",
        "required": false,
        "schema": {
          "type": "string",
          "format": "datetime"
        }
      },
      "filter.updatedAtAfter": {
        "name": "updatedAt[after]",
        "in": "query",
        "description": "Updated after the provided date",
        "required": false,
        "schema": {
          "type": "string",
          "format": "datetime"
        }
      }
    },
    "schemas": {
...
      "input.barometer": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "The barometer type",
            "example": "happiness",
            "pattern": "^\\w+$"
          },
          "score": {
            "type": "integer",
            "description": "score of the barometer",
            "example": 50,
            "minLength": 1,
            "maxLength": 100
          }
        },
        "required": [
          "type",
          "score"
        ]
      },
...
      "output.message": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "details"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "output.barometers": {
        "type": "array",
        "description": "Barometers list",
        "items": {
          "$ref": "#/components/schemas/output.barometer"
        }
      },
      "output.barometer": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "The barometer type",
            "example": "happiness"
          },
          "score": {
            "type": "integer",
            "description": "Barometer score value",
            "example": 50,
            "minimum": 1,
            "maximum": 100
          },
          "date": {
            "type": "string",
            "format": "date",
            "example": "2022-01-01",
            "description": "The barometer date"
          }
        },
        "required": [
          "type",
          "score"
        ]
      },
      "output.barometer.stats": {
        "$ref": "#/components/schemas/output.barometer.stats.content"
      },
      "output.barometer.stats.content": {
        "type": "array",
        "items": {
          "required": [
            "score",
            "type",
            "year",
            "entries"
          ],
          "type": "object",
          "properties": {
            "score": {
              "type": "number",
              "description": "The average score value for the requested period",
              "example": 10
            },
            "type": {
              "type": "string",
              "example": "foo",
              "description": "The barometer name (it's type)"
            },
            "workCategoryId": {
              "type": "integer",
              "example": 1,
              "description": "The associated work category of user."
            },
            "year": {
              "type": "integer",
              "example": 2022,
              "description": "Year of the stat range (maximal range)."
            },
            "month": {
              "type": "integer",
              "example": 5,
              "description": "Month of the stat range, optional."
            },
            "day": {
              "type": "integer",
              "example": 12,
              "description": "The day of the stat range (mininal range)"
            },
            "entries": {
              "type": "integer",
              "example": 500,
              "description": "Number of barometers filled for this period and scope"
            }
          }
        }
      },
      "output.notifications": {
        "type": "array",
        "description": "Notifications list",
        "items": {
          "$ref": "#/components/schemas/output.notification"
        }
      },
      "output.notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1,
            "description": "The user account id"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "read": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "New notification",
            "nullable": false
          },
          "level": {
            "type": "integer",
            "description": "The notification priority level.",
            "minimum": 1,
            "maximum": 3
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "read",
          "message",
          "level"
        ]
      },
...
    }
  }
}